% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello(name = "World")
}
\arguments{
\item{name}{A string, the user's name}
}
\value{
A greeting message
}
\description{
This function says hello to the user.
}
\examples{
hello("Alice")
}
